move plugins folder and add error handling
