+++ src/plugins/cache.ts
@@ -1,10 +1,12 @@
 import {setup} from "axios-cache-adapter";
-import {Clockodo} from "../api";
-import {axiosClient} from "../utilities/symbols";
+import {Clockodo} from "../internals/api";
+import {axiosClient} from "../internals/utilities/symbols";
 
 const cachePlugin = (config: {cacheTime: number}) => (clockodo: Clockodo) => {
-    if (typeof config.cacheTime === "number") {
-        clockodo[axiosClient] =
+    if (typeof config.cacheTime !== "number") {
+        throw new Error("Clockodo cacheTime expected to be a number, is typeof: " + typeof config.cacheTime);
+    }
+    clockodo[axiosClient] =
         setup({
             baseURL: clockodo[axiosClient].defaults.baseURL,
             headers: clockodo[axiosClient].defaults.headers,
@@ -13,7 +15,6 @@ const cachePlugin = (config: {cacheTime: number}) => (clockodo: Clockodo) => {
                 exclude: {query: false},
             },
         });
-    }
 };
 
 export default cachePlugin;
