+++ tests/unit/unit.test.ts
@@ -9,7 +9,7 @@ import {
     ENTRY_BILLED,
     ABSENCE_TYPE_SPECIAL_LEAVE,
 } from "../../src/internals/api";
-import cachePlugin from "../../src/internals/plugins/cache";
+import cachePlugin from "../../src/plugins/cache";
 
 const CLOCKODO_API = "https://my.clockodo.com/api";
